feat: Add user input verification, code cleanup
This commit will add / refactor the user input verfication of the
town / address via two dedicated functions. It should cover most
of the possible address formats used for a town. This is important
to protect the geopy api from overload.

This commit will correct the style of the python code in order
to be inline with the python code style check.

This commit will also contain a draft version of stitching together
the return value back to the user, with the weather data.