feat: Activate MeteoDataCall class & add testcode
This commit will basically activate the MeteoDataCall class and
call the live_data function out of it. It will also include the
error handling for the open-weather api and add some testcode to
be quicker when testing. It should be prevented to stress the
geopy api (disclaimer in the use policy of the api) and therfore
some local data is added.

- Activate MeteoDataCall class
- Start using the live_data method out of the class
- Add error handling for weather api
- Add some local testing data and testing code