feat: Add my_emoji file with emoji hash
This commit will add (rename emoji.py) the my_emoji.py file to the
repo, which contains a hash of emoji's. The key value of each
emoji correlates with the name of the weatherconditions in the
translate_weathercode() function. Therefore a correlation of the
weathercondidtion to the key of the emoji can be automatically be
done.