feat: Add function to loop live weather
This commit will add the live_weather_loop function, which
repeatedly asks the user if he wants to get the live weather for
another town. If yes, the loop continues, if no it will break and
bring the user back to the main menu, which will call the main()
function. (The main() function is not present yet in the code)