feat: Add error handling for geopy API call
This commit will add enhanced error handling for the geopy module / API
call and make sure, all possible errors are handled for each of the
various requests.
It will also verify if the geopy location lookup will return anything
at all or not. If not, an error is displayed and the user is sent
back to main menu, the weather-api will NOT happen since there is
no proper geodata.