feat: Improve error handling
This commit will improve the error handling for the MeteoDataCall
class, especially the live_data method in it. It will handle known
http errors from the open-weather api as well as any unknown error
of it. In both cases, the program will exit, since there is not reason
to continue the program, if the api returns an error.

It will also improve the get_live_weather function for it's error
handling, when calling the geopy api. If geopy returns None, it will
also terminate the program, no reason to continue.
If the api returns something "strange" and the coordiantes array
includes invalid data, whith which the open-weather api cannot deal,
the error handling of the open-weather api will handle it.