feat: Enhance input validation
This commit will add enhanced input validation for the repetitive
questions, where the user has to answer with Y/N. It will fix the
problem, that if a user inputs anything else than Y/N, the loop for
live weatehr or historical weather would still continue.
This was due to the fact, that the input verification AND the function
call was in the while loop.

The input validation is now distributed in a separate function, which
is generic. It takes the question and the possible answers as argument
and until the desired answer is provided as input, it will loop/repeat.

Due to this change, the main function had to be refactored, the menu
questionary had to be moved into the while loop.