docs: Add various chapters to the testing & readme
This commit will move some content from the README to the TESTING file,
since that is the place where it belongs (part of error handling
testing)
It will also add the Error Handling Testing chapter as well as the
Python breakpoint section.
Some general cleanup of the TESTING file is done.