main: Cleanup code and comments
This commit will do some code cleanup. It will remove the main
"if / else" condition form live_weatehr and historical_weather fuction.
This condition wass an old code block required for
the error handling, before I had separate classes for the api call.
The code was now for nothing and did not make sense, therefore I did
remove it.

The second part of the commit will add/correct a lot of comments or
description parts in the code.