refactor: Improve main function input validation
This commit will improve the validation of the user input for the main
menu. It will also use the external ask_input() function. Therefore
the validation of the answer is done on that external function and
the "else" statement in the condition of the main menu, can be removed.
When giving false inputs, this else did anyway not work as expected.